backend/requirements.txt
# DNSMate Backend Dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication
fastapi-users[sqlalchemy]==12.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database (SQLAlchemy 2.0+ for compatibility with fastapi-users)
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Configuration
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2

# Form data handling
python-multipart==0.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0

# Development tools
watchdog==3.0.0
ipdb==0.13.13